pyproject.toml
[tool.poetry]
name = "mobilegymapp"
version = "0.1.0"
description = ""
authors = ["xMati278 <59803172+xMati278@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
kivy = "^2.3.0"
kivymd = "^1.2.0"
pre-commit = "^3.6.2"
flake8 = "^7.0.0"
pytest = "^8.0.2"
bandit = "^1.7.7"
safety = "^3.0.1"
ruff = "^0.3.0"
python-dotenv = "^1.0.1"
icecream = "^2.1.3"
python-for-android = "^2024.1.21"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
